requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
atlassian-python-api==3.14.0
attrs==21.2.0
beautifulsoup4==4.10.0
bs4==0.0.1
cached-property==1.5.2
certifi==2021.5.30
chardet==3.0.4
click==7.1.2
Deprecated==1.2.13
expiringdict==1.2.1
gTTS==2.2.1
idna==2.10
lxml==4.6.3
mailru-im-bot==0.0.19
monotonic==1.5
multidict==5.1.0
oauthlib==3.1.1
python-baseconv==1.2.2
requests==2.25.0
requests-oauthlib==1.3.0
six==1.15.0
soupsieve==2.2.1
termcolor==1.1.0
terminaltables==3.1.0
typing-extensions==3.10.0.2
urllib3==1.26.3
wrapt==1.12.1
yarl==1.6.3
